Move code unused by prod but maybe useful in test
